PG-adapting: delete some outdated prooftree information
Delete functions from PG-adapting that have been deleted in the
sources to fix make magic. Delete also some of the outdated
information of the Prooftree configuration section. Note that
this section is still outdated, which is hopefully OK, given the
number of attempts to add Prooftree support for more proof
assistants in the last years.